+++ app/Models/RaterGroup.php
@@ -2,11 +2,14 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class RaterGroup extends Model
 {
+    use HasFactory;
+
     protected $fillable = [
         'subject_id',
         'name',